config class moved to servicebus package, verify added to test, code tidy up
