+++ src/main/resources/application.yaml
@@ -181,7 +181,7 @@ dbFeature:
   flags:
     enable: ${DB_FEATURE_FLAG_ENABLE:}
     disable: ${DB_FEATURE_FLAG_DISABLE:}
-logging.level.uk.gov.hmcts: DEBUG
+logging.level.uk.gov.hmcts: ${APPLICATION_LOGGING_LEVEL:INFO}
 
 springdoc:
   swagger-ui:
